# Optimize protobuf encoder and iterator for various workloads