--- a/src/dbnode/encoding/m3tsz/float_encoder_iterator.go
+++ b/src/dbnode/encoding/m3tsz/float_encoder_iterator.go
@@ -26,6 +26,11 @@ import (
 	"github.com/m3db/m3/src/dbnode/encoding"
 )
 
+const (
+	bits12To6Mask = 4032 // 1111 1100 0000
+	bits6To0Mask  = 63   // 0011 1111
+)
+
 // FloatEncoderAndIterator encapsulates the state required for a logical stream of bits
 // that represent a stream of float values compressed with XOR.
 type FloatEncoderAndIterator struct {
@@ -145,8 +150,8 @@ func (eit *FloatEncoderAndIterator) readNextFloat(stream encoding.IStream) error
 		return err
 	}
 
-	numLeadingZeros := (numLeadingZeroesAndNumMeaningfulBits & 4032) >> 6
-	numMeaningfulBits := (numLeadingZeroesAndNumMeaningfulBits & 63) + 1
+	numLeadingZeros := (numLeadingZeroesAndNumMeaningfulBits & bits12To6Mask) >> 6
+	numMeaningfulBits := (numLeadingZeroesAndNumMeaningfulBits & bits6To0Mask) + 1
 
 	meaningfulBits, err := stream.ReadBits(int(numMeaningfulBits))
 	if err != nil {