--- a/src/dbnode/encoding/proto/common.go
+++ b/src/dbnode/encoding/proto/common.go
@@ -153,21 +153,18 @@ type sortedMarshalledFields []marshalledField
 // customFieldState is used to track any required state for encoding / decoding a single
 // field in the encoder / iterator respectively.
 type customFieldState struct {
-	fieldNum       int
-	fieldType      customFieldType
-	protoFieldType dpb.FieldDescriptorProto_Type
-
+	fieldNum int
 	// Float state. Works as both an encoder and iterator (I.E the encoder calls
 	// the encode methods and the iterator calls the read methods).
 	floatEncAndIter m3tsz.FloatEncoderAndIterator
-
 	// Bytes State. TODO(rartoul): Wrap this up in an encoderAndIterator like
 	// the floats and ints.
 	bytesFieldDict         []encoderBytesFieldDictState
 	iteratorBytesFieldDict [][]byte
-
 	// Int state.
-	intEncAndIter intEncoderAndIterator
+	intEncAndIter  intEncoderAndIterator
+	protoFieldType dpb.FieldDescriptorProto_Type
+	fieldType      customFieldType
 }
 
 type encoderBytesFieldDictState struct {