--- a/src/dbnode/encoding/proto/custom_marshal.go
+++ b/src/dbnode/encoding/proto/custom_marshal.go
@@ -30,8 +30,12 @@ type customFieldMarshaler interface {
 	encFloat64(tag int32, x float64)
 	encFloat32(tag int32, x float32)
 	encInt32(tag int32, x int32)
+	encSInt32(tag int32, x int32)
+	encSFixedInt32(tax int32, x int32)
 	encUInt32(tag int32, x uint32)
 	encInt64(tag int32, x int64)
+	encSInt64(tag int32, x int64)
+	encSFixedInt64(tax int32, x int64)
 	encUInt64(tag int32, x uint64)
 	encBool(tag int32, x bool)
 	encBytes(tag int32, x []byte)
@@ -51,25 +55,45 @@ func newCustomMarshaler() customFieldMarshaler {
 }
 
 func (m *customMarshaler) encFloat64(tag int32, x float64) {
+	if x == 0 {
+		// Default values are not included in the stream.
+		return
+	}
+
 	m.buf.encodeTagAndWireType(tag, proto.WireFixed64)
 	m.buf.encodeFixed64(math.Float64bits(x))
 }
 
 func (m *customMarshaler) encFloat32(tag int32, x float32) {
+	if x == 0 {
+		// Default values are not included in the stream.
+		return
+	}
+
 	m.buf.encodeTagAndWireType(tag, proto.WireFixed32)
 	m.buf.encodeFixed32(math.Float32bits(x))
 }
 
 func (m *customMarshaler) encBool(tag int32, x bool) {
-	var val uint64
-	if x {
-		val = 1
+	if !x {
+		// Default values are not included in the stream.
+		return
 	}
-	m.encUInt64(tag, val)
+
+	m.encUInt64(tag, 1)
 }
 
 func (m *customMarshaler) encInt32(tag int32, x int32) {
 	m.encUInt64(tag, uint64(x))
+}
+
+func (m *customMarshaler) encSInt32(tag int32, x int32) {
+	m.encUInt64(tag, encodeZigZag32(x))
+}
+
+func (m *customMarshaler) encSFixedInt32(tag int32, x int32) {
+	m.buf.encodeTagAndWireType(tag, proto.WireFixed32)
+	m.buf.encodeFixed32(uint32(x))
 }
 
 func (m *customMarshaler) encUInt32(tag int32, x uint32) {
@@ -80,12 +104,31 @@ func (m *customMarshaler) encInt64(tag int32, x int64) {
 	m.encUInt64(tag, uint64(x))
 }
 
+func (m *customMarshaler) encSInt64(tag int32, x int64) {
+	m.encUInt64(tag, encodeZigZag64(x))
+}
+
+func (m *customMarshaler) encSFixedInt64(tag int32, x int64) {
+	m.buf.encodeTagAndWireType(tag, proto.WireFixed64)
+	m.buf.encodeFixed64(uint64(x))
+}
+
 func (m *customMarshaler) encUInt64(tag int32, x uint64) {
+	if x == 0 {
+		// Default values are not included in the stream.
+		return
+	}
+
 	m.buf.encodeTagAndWireType(tag, proto.WireVarint)
 	m.buf.encodeVarint(x)
 }
 
 func (m *customMarshaler) encBytes(tag int32, x []byte) {
+	if len(x) == 0 {
+		// Default values are not included in the stream.
+		return
+	}
+
 	m.buf.encodeTagAndWireType(tag, proto.WireBytes)
 	m.buf.encodeRawBytes(x)
 }
@@ -99,5 +142,9 @@ func (m *customMarshaler) bytes() []byte {
 }
 
 func (m *customMarshaler) reset() {
-	m.buf.reset(m.buf.buf)
+	b := m.buf.buf
+	if b != nil {
+		b = b[:0]
+	}
+	m.buf.reset(b)
 }