… src/dbnode/encoding/ostream_test.go
… src/dbnode/encoding/proto/benchmarks_test.go
… src/dbnode/encoding/proto/custom_marshal_test.go
… src/dbnode/encoding/proto/custom_unmarshal_test.go
… src/dbnode/encoding/proto/docs/README.md
… src/dbnode/encoding/proto/docs/encoding.md
… src/dbnode/encoding/proto/docs/marshal.md
… src/dbnode/encoding/proto/docs/unmarshal.md
… src/dbnode/encoding/proto/encoder.go
… src/dbnode/encoding/proto/encoder_test.go
… src/dbnode/encoding/proto/round_trip_prop_test.go
… src/dbnode/encoding/proto/round_trip_test.go
… src/dbnode/namespace/schema.go
… src/m3ninx/postings/pilosa/codec_test.go
… src/msg/protocol/proto/types.go