**What this PR does / why we need it**:
Remove `jhump/protoreflect` from the hot path as much as possible, plus some other optimizations and cleanup.

Results:

1. 10x improvement in encoding and decoding for non-custom fields.
2. 2x improvement in encoding for custom fields.
3. 4x improvement in decoding for custom fields (original goal).

before
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/m3/src/dbnode/encoding/proto
BenchmarkEncoder/with_non_custom_encoded_fields_enabled-8         	    1000	   1432754 ns/op	  374435 B/op	    8177 allocs/op
BenchmarkEncoder/with_non_custom_encoded_fields_disabled-8        	   10000	    111547 ns/op	   13430 B/op	       0 allocs/op
BenchmarkIterator/with_non_custom_encoded_fields_enabled-8        	    1000	   1274402 ns/op	  323059 B/op	    8196 allocs/op
BenchmarkIterator/with_non_custom_encoded_fields_disabled-8       	   10000	    203996 ns/op	   21603 B/op	     910 allocs/op
```

after
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/m3/src/dbnode/encoding/proto
BenchmarkEncoder/with_non_custom_encoded_fields_enabled-8         	   10000	    119435 ns/op	   66221 B/op	       0 allocs/op
BenchmarkEncoder/with_non_custom_encoded_fields_disabled-8        	   20000	     67853 ns/op	   21176 B/op	       0 allocs/op
BenchmarkIterator/with_non_custom_encoded_fields_enabled-8        	   20000	     90630 ns/op	     671 B/op	      13 allocs/op
BenchmarkIterator/with_non_custom_encoded_fields_disabled-8       	   30000	     46309 ns/op	     595 B/op	      12 allocs/op
```